refactor: extract BaseCNN to separate module and resolve CodeScene issues
Resolves three CodeScene quality gate issues:

- Low Cohesion in kale/embed/cnn.py

- Complex Method in BaseCNN._create_sequential_conv_blocks

- Low Cohesion in tests/embed/test_cnn.py

Changes:

1. Extract BaseCNN utility class to kale/embed/base_cnn.py (136 statements)

   - Centralizes CNN building utilities and helper methods

   - Reduces cnn.py from 904 to 86 lines (90% reduction)

   - Clear separation: base_cnn.py (utilities) vs cnn.py (architectures)

2. Reduce complexity of _create_sequential_conv_blocks (17 -> 3)

   - Extract _validate_conv_block_inputs() for input validation

   - Extract _normalize_conv_parameters() for parameter normalization

   - Extract _validate_parameter_lengths() for length checking

3. Split test file to mirror module structure

   - Create tests/embed/test_base_cnn.py (37 BaseCNN utility tests)

   - Reduce tests/embed/test_cnn.py (18 concrete architecture tests)

   - Test structure now matches production code organization

4. Update imports across dependent modules

   - kale/embed/image_cnn.py: import BaseCNN from base_cnn

   - kale/embed/signal_cnn.py: import BaseCNN from base_cnn

   - Direct imports (no re-exports) for cleaner architecture

Test Results:

- All 100 tests passing (55 split + 45 image_cnn + 6 signal_cnn)

- 100% coverage maintained for base_cnn.py and cnn.py

- Zero regressions in functionality
